clk: rockchip: introduce pvtpll-adaptive driver
this driver can adaptively tune in the requested frequncy by carefully
walking through the given ring length window.

It uses status grf pvtpll status registers to measure the actual pvtpll
frequency and increase the ring lenght if frequency needs to be lowered
or decrease the ring length if the frequency needs to be increased.

During the frequency seaarch, if the target is overshot or undershot, it
uses the ring length which generates the closest frequency to the target
rate.

When configuring the grf, the pd and its related clocks must be gated
on. To achieve it pvtpll manages the lifecycle of the clocks explicity
when active together with the power domain so that consumer driver dont
need to care such clock quirks during muxing transitions.

All registers are configurable through dts and additionaly can read the
OTP for vendor guranteed operating frequency window.