OPP: Handle CLK_VOLTAGE_FIRST
Specific clock types (currently PVTPLL based clocks used in Rockchip
SOCs) needs to be voltaged before regardless they are being upscaled or
downscaled.

This commit checks the clock topology and if there is a clock marked as
CLK_VOLTAGE_FIRST flag and sets the voltage first acording to the
availability.

This commit is only a draft for RFC, there might be further
necessesities to check ie: CLK_SET_RATE_PARENT.

One might argue that such oddities must be configured in config_clks
callbacks, but having this done causes that the drivers using such
clocks always need to implement platform specific code into it which
causes driver to be platform specific and not generic. Example: panthor.